better support for maven version strings
Unless the maven version string is a valid Gem 
version string create a substitute gem version 
string by dividing the maven version string into 
it's numeric elements and joining them back 
together with '.' characters.

The string 'alpha' in a maven version string 
is converted to '0'.

The string 'beta' in a maven version string 
is converted to '1'.

In general gem versions strings need to either 
be an Integer or start with a digit and a '.'.

For example the flying saucer core-renderer jar 
that uses itext to generate a pdf from styled 
xhtml input has a version string of "R8pre2". 

Installing this jar:

  jruby -S gem maven org/xhtmlrenderer core-renderer R8pre2

now results in:

  Successfully installed core-renderer-8.2-java
  1 gem installed

  jruby -S gem list core-renderer
  
  *** LOCAL GEMS ***
  
  core-renderer (8.2)

In addition the following constants are created in 
the new maven gem:

  CoreRenderer::VERSION         # => "8.2"
  CoreRenderer::MAVEN_VERSION   # => "R8pre2"

Signed-off-by: Charles Oliver Nutter <charles.nutter@sun.com>